// src/app/model/Mensualidad.ts
export class Mensualidad {

    id: number;
    numero_contrato: number;
    objeto_contrato: string;
    nombreContratista: string;
    ruc_contratista: number;
    numOrden: number;
    cantidad: number;
    precioUnitario: number;
    subtotal: number;
    iva: number;
    otrosImpuestos: number;
    total: number;
    cod_presupuestario: number;
    descripcionpresupuestaria: string;
    descripcion: string;
    valorCert: number;
    afectuar: number;
    saldo: number;
    fechaHoraIni: string;
    email: string;
    obtencion_iva: number;
    obtencion_otro_impuesto: number;
    meses_acumulables: number;
    tipo_pago: string;
    mes: string;
    active:boolean;
}